requirements.txt
alembic
fastapi
mysqlclient
passlib[bcrypt]
python-jose[cryptography]
python-multipart
sqlalchemy
uvicorn[standard]